#188 - move images extractor to extractors dir and correct videos
