+++ goose/extractors/videos.py
@@ -22,7 +22,7 @@ limitations under the License.
 """
 
 from goose.extractors import BaseExtractor
-from goose.videos import Video
+from goose.video import Video
 
 VIDEOS_TAGS = ['iframe', 'embed', 'object', 'video']
 VIDEO_PROVIDERS = ['youtube', 'vimeo', 'dailymotion', 'kewego']
