+++ goose/image.py
