+++ goose/utils/images.py
@@ -25,8 +25,8 @@ import os
 import urllib2
 from PIL import Image
 from goose.utils.encoding import smart_str
-from goose.images.image import ImageDetails
-from goose.images.image import LocallyStoredImage
+from goose.image import ImageDetails
+from goose.image import LocallyStoredImage
 
 
 class ImageUtils(object):
