+++ goose/video.py
